test: add regex escape tests and mongo test refactor
